Fix medium/high severity bugs from ingredient standardization audit.
- B2 (High): searchByVariants() now queries all variants via .or() instead of only variants[0]
- B6 (Medium): Raise OpenAI max_tokens 1000→4096 to prevent silent batch truncation to confidence:0
- B1 (Medium): getCanonicalNameSample() orders by id DESC so recent canonicals appear in AI context
- B3 (Medium): Extract NON_FOOD_TITLE_TOKENS/PHRASES/hasNonFoodTitleSignals to shared module; removes triplication across 3 files
- B4 (Medium): Move backfillRecipeIngredientReviewFlags() from hot claimPending() path to one-time call at pipeline startup
- B5 (High): Default allowFallback to false in claimPending() so the racy non-atomic fallback is never used in production
