+++ backend/workers/standardizer-worker/realtime-standardizer.ts
@@ -1,51 +1,7 @@
 import { normalizeCanonicalName, singularizeCanonicalName } from "../../scripts/utils/canonical-matching"
 import type { IngredientStandardizerContext } from "./ingredient-standardizer"
 import type { IngredientStandardizationInput, IngredientStandardizationResult } from "./ingredient-standardizer"
-
-const NON_FOOD_TITLE_TOKENS = new Set([
-  "balm",
-  "body",
-  "candle",
-  "conditioner",
-  "cosmetic",
-  "deodorant",
-  "dog",
-  "face",
-  "fragrance",
-  "lotion",
-  "lip",
-  "makeup",
-  "mask",
-  "perfume",
-  "pet",
-  "shampoo",
-  "skincare",
-  "soap",
-  "scented",
-  "toothpaste",
-  "toy",
-  "treat",
-  "treats",
-  "cat",
-  "litter",
-])
-
-const NON_FOOD_TITLE_PHRASES = [
-  ["body", "butter"],
-  ["body", "oil"],
-  ["body", "wash"],
-  ["face", "mask"],
-  ["lip", "balm"],
-  ["lip", "mask"],
-  ["lip", "oil"],
-  ["lip", "gloss"],
-  ["pet", "treats"],
-  ["dog", "treats"],
-  ["cat", "treats"],
-  ["dog", "food"],
-  ["cat", "food"],
-  ["tooth", "paste"],
-]
+import { hasNonFoodTitleSignals } from "../shared/non-food-signals"
 
 const PROTECTED_FORM_TOKENS = new Set([
   "seed",
@@ -138,21 +94,6 @@ const OPTIONAL_PHRASE_RE = /\b(?:to taste|if needed|as needed|optional|divided)\
 
 const TRAILING_PACKAGING_RE =
   /\s+\d+(?:\.\d+)?\s*(?:oz|ounce|ounces|lb|lbs|pound|pounds|g|gram|grams|kg|kilogram|kilograms|ml|milliliter|milliliters|l|liter|liters|litre|litres|ltr|qt|quart|quarts|pt|pint|pints|gal|gallon|gallons|fl\s*oz|floz|ct|count|counts|pk|pkg|pack|packs|package|packages|bottle|bottles|bag|bags|box|boxes|can|cans|jar|jars|carton|cartons|tray|trays|case|cases|pouch|pouches|unit|units|each|ea|piece|pieces)\b.*$/i
-
-function hasNonFoodTitleSignals(sourceName: string): boolean {
-  const normalized = normalizeCanonicalName(sourceName)
-  if (!normalized) return false
-
-  const tokens = normalized.split(" ").filter(Boolean)
-  if (!tokens.length) return false
-
-  const tokenSet = new Set(tokens)
-  if (tokens.some((token) => NON_FOOD_TITLE_TOKENS.has(token))) {
-    return true
-  }
-
-  return NON_FOOD_TITLE_PHRASES.some((phrase) => phrase.every((token) => tokenSet.has(token)))
-}
 
 function cleanIngredientName(name: string): string {
   return normalizeCanonicalName(name)
